server_django/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
python-dotenv>=1.0,<1.1
mysqlclient>=2.2,<2.3
cryptography>=42.0,<43.0
gunicorn>=21.2,<22.0
